Remove the `require-await` rule
It caused way too many false-positives.

Just because an async function doesn't have `await`, doesn't mean it can be removed without changing the behavior. The function could have a method call that could throw and then removing the `async` keyword changes the behavior as it now suddenly throws instead of rejecting. We could maybe detect all `throw` statements directly in the async function body, but there might be function calls that throw many levels deep. It's just impossible to reliably detect this. This rule has been on for a while now and I constantly have to `eslint-ignore` valid code.

Even for async functions that just return a promise and doesn't await, the async keyword can be a nice indicator for humans reading the code about what it returns.